# internal/common/infra/api/swagger/swagger.yaml
basePath: /v1
definitions:
  schemas.ConfirmPaymentSchema:
    properties:
      action:
        type: string
      api_version:
        type: string
      data:
        properties:
          id:
            type: string
        type: object
      date_created:
        type: string
      id:
        type: string
      live_mode:
        type: boolean
      type:
        type: string
      user_id:
        type: integer
    type: object
  schemas.CreatePixBillingSchema:
    properties:
      amount:
        type: number
      order_id:
        type: string
    type: object
  schemas.PaymentResponseSchema:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      method:
        type: string
      order_id:
        type: string
      paid_at:
        type: string
      qr_code:
        type: string
      status:
        type: string
      transaction_code:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an API for a payments on tech challenge.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: API Microservice Payment
  version: "1.0"
paths:
  /payments/order/{orderID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PaymentResponseSchema'
        "500":
          description: Internal Server Error
      summary: Get a payment by OrderID
      tags:
      - Payments
  /payments/pix:
    post:
      consumes:
      - application/json
      parameters:
      - description: Pix billing to create
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/schemas.CreatePixBillingSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.PaymentResponseSchema'
        "500":
          description: Internal Server Error
      summary: Create Pix Billing
      tags:
      - Payments
  /payments/webhook:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment confirmation webhook
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/schemas.ConfirmPaymentSchema'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
      summary: Confirm Payment
      tags:
      - Payments
schemes:
- http
swagger: "2.0"